Updates the symbolic methods
Moves all methods to the symbolic instructions package. By this the methods can access symbolic state.

Improves the selectState method. Because the states are split based on their choice PC no longer needs to be updated.

Improves performance of the newSymbol method. Caches the symbols so that they do not need to be constructed per vm instance.

Makes all random selection methods use the `pushValues` method.